remove redundant file and inherit from the original
